finished autotagger. larger seed
so the newsapi was paginated this entire time..... some things with the api dont work as expected (namely AND/ORing topics) but this should work for now.
the new seeder takes the first 5 pages from the new api, which then causes the seeder to run for quite some time. we can discuss cutting that down later.